Add a status field in the Attendance command functionality
To provide a more comprehensive means attendance, we must put
into consideration the types of student attendance. There's
present, absent and absence due to valid reason (VR).

Of course, this can be an issue if our current application
only supports a mere "marked/unmark" status for attendance,
should the circumstances press the user (a TA) to verify
attendance records.

So let's add a new field to the attendance command!

Added support for attendance statuses.

In addition, all references to "AddressBook" has been
replaced with "NpcTrack".

Added a new logo as well!