Amend error detection behaviour in MarkAttendanceParser
The MarkAttendanceParser.java classes was not throwing the correct error message for
commands with missing parameters or incorrect formats.

Let's ammend MarkAttendanceParser.java to throw the correct error for missing parameters
and incorrect formats!

Also updated my documentation of contribution to be more consistent with the rest of
the team's contribution documents.